# backend/server/openapi-clean.yaml
openapi: 3.0.3
info:
  title: ContractSecure API
  version: "1.0.0-mvp"
  description: |
    Secure digital contract platform: create contracts, invite signers, verify identity via OTP,
    capture consent + GPS, embed signatures, finalize tamper-proof PDF with SHA-256.
  contact:
    name: API Support
    email: support@contractsecure.example.com
  license:
    name: Proprietary
    url: https://contractsecure.example.com/terms
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.contractsecure.example.com
    description: Production
...
paths:
...
  /contracts:
    post:
      tags: [Contracts]
      summary: Create a new contract
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, file]
              properties:
                title: { type: string, example: "Employment Agreement" }
                file: { type: string, format: binary }
      responses:
        '201':
          description: Contract created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /contracts/{contractId}:
    get:
      tags: [Contracts]
      summary: Get contract details
      parameters:
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: Contract details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '404':
          $ref: '#/components/responses/NotFound'

  # Signers
...
  /contracts/{contractId}/download:
    get:
      tags: [Files]
      summary: Download the final signed contract
      parameters:
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    ContractId:
      name: contractId
      in: path
      required: true
      schema: { type: string }
      description: Contract ID
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Contract:
      type: object
      properties:
        id: { type: string, example: "contract_12345" }
        title: { type: string, example: "Employment Agreement" }
        status: { $ref: "#/components/schemas/ContractStatus" }
        created_at: { type: string, format: date-time, nullable: true }
        base_pdf_path: { type: string, nullable: true }
        final_pdf_path: { type: string, nullable: true }
        pdf_sha256: { type: string, nullable: true }
    ContractStatus:
      type: string
      enum: [draft, pending, partially_signed, executed, void, expired, declined]
      description: |
        - draft: Contract is being prepared
        - pending: Waiting for signatures
        - partially_signed: Some but not all signatures collected
        - executed: Fully signed and completed
        - void: Cancelled/voided
        - expired: Signing period expired
...
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error: { type: string, example: "not_found" }
        message: { type: string, example: "The requested resource was not found" }
        details: 
          type: array
          items:
            type: object
            properties:
              field: { type: string, example: "email" }
              message: { type: string, example: "Invalid email format" }
        requestId: { type: string, format: uuid, example: "550e8400-e29b-41d4-a716-446655440000" }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...